feat(workspace): 파일 업로드 및 자동 라벨링 기능 작업
- 워크스페이스 생성 시 파일 업로드 지원
- 자동 라벨링 페이지에서 워크스페이스 선택 시 해당 파일 조회 및 라벨링 기능 활성화
- 이미지 카드 및 라벨별 이미지 그룹화 컴포넌트 추가
- 결과 상세 페이지에서 라벨별 이미지 카드로 변경하여 UI 개선

+++ UserServer/src/main/java/cc/nobrain/dev/userserver/domain/train/service/dto/TrainRes.kt
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor
 
 data class LabelAndIds (
     val label: String,
-    val ids: List<Int>
+    val ids: List<Long>
 )
 
 data class TrainRes(
